split up functions to separate files for modularity
