+++ rust-implementation/rust-engine/src/moves/magic/mod.rs
@@ -1,2 +1,9 @@
-mod magic;
-pub use magic::*;
+pub mod attacks;
+pub mod lookup;
+pub mod masks;
+pub mod search;
+
+pub use attacks::*;
+pub use lookup::*;
+pub use masks::*;
+pub use search::*;
